add a clear history/cookies options
when deleting media from album while user is actively viewing the media it doesnt change (same issue when moving )
made sure when the user deletes it will also update the windowedlist and windowedlistindex 

fix referencing nsmanaged objects after they are deleted in asyncphotoimage causing a breakpoint error

remove media reference from select media entity

start rewrite of fullcoversheet to remove useless and bad code

readded the onChange for windowList

implement rest of fullcoversheet

re implemented all of the bottom header buttons 

added better toast messaging for tracking export progress 

instead of tracking 2 indexes (windowListIndex and mediaListIndex) I created a single point of truth
where i am able to compute those indexes from currentMediaId: UUID 

fixed error with exporting media happened twice and added better ui updates for exporting single item

implement better error displaying for fullcoversheet

added a toast viewmodifer and view to handle displaying toast messages instead of using same overlay code 

fixed error handling in exporting video to show all failures via toast 

rewrote the export photo func to check if user has given access to library and display proper toast message if not 

implement error displaying for full cover sheet

+++ PhotoSafe/Services/MediaService.swift
@@ -11,6 +11,7 @@ import CoreData
 enum MediaError: Error {
     case invalidImageData
     case invalidUrl
+    case failedFetch
 }
 
 // Define the blueprint for AlbumService
@@ -18,23 +19,66 @@ protocol MediaServiceProtocol {
     func save_media(to album: AlbumEntity, type: MediaType, imageData: Data, thumbnail: Data, videoPath: String?) throws -> MediaEntity
     func fetch_media(from album: AlbumEntity) -> [MediaEntity]
     func fetchAll() -> [MediaEntity]
-    func delete(media: MediaEntity) throws
-    func move(media: MediaEntity, to album: AlbumEntity) throws
-    func like_or_unlike(with status: Bool, for media: MediaEntity) throws -> MediaEntity
+    func delete(id: UUID) throws
+    func move(id: UUID, to album: AlbumEntity) throws
+    func fetchFavorites() -> [MediaEntity]
+    func favorite(for id: UUID) throws -> MediaEntity
+    func unfavorite(for id: UUID) throws -> MediaEntity
+    //func fetchById(id: UUID) throws -> MediaEntity?
 }
 
 final class MediaService: MediaServiceProtocol {
+    private func fetchById(id: UUID) throws -> MediaEntity? {
+        let request = MediaEntity.fetchRequest()
+        request.predicate = NSPredicate(format: "id == %@", id as CVarArg)
+        request.fetchLimit = 1
+        guard let media = try? context.fetch(request).first else { return nil }
+        return media
+    }
+    
+    func favorite(for id: UUID) throws -> MediaEntity {
+        do {
+            guard let mediaEntity = try self.fetchById(id: id) else { throw MediaError.failedFetch }
+            
+            mediaEntity.is_favorited = true
+            try self.context.save()
+            return mediaEntity
+        } catch (let error) {
+            throw error
+        }
+    }
+    
+    func unfavorite(for id: UUID) throws -> MediaEntity {
+        do {
+            guard let mediaEntity = try self.fetchById(id: id) else { throw MediaError.failedFetch }
+            
+            mediaEntity.is_favorited = false
+            try self.context.save()
+            return mediaEntity
+        } catch (let error) {
+            throw error
+        }
+    }
+    
     private let context: NSManagedObjectContext
     
     init(context: NSManagedObjectContext = CoreDataManager.shared.container.viewContext) {
         self.context = context
     }
     
-    func like_or_unlike(with status: Bool, for media: MediaEntity) throws -> MediaEntity {
-        media.is_favorited = status
-        try self.context.save()
-        return media
-    }
+//    func like_or_unlike(with status: Bool, for id: UUID) throws -> MediaEntity {
+//        do {
+//            guard let media = try self.context.fetch(MediaEntity.fetchRequest()).first(where: { $0.id == id }) else {
+//                throw MediaError.failedFetch
+//            }
+//            
+//            media.is_favorited = status
+//            try self.context.save()
+//            return media
+//        } catch (let error) {
+//            throw error
+//        }
+//    }
     
     func save_media(
         to album: AlbumEntity,
@@ -73,16 +117,24 @@ final class MediaService: MediaServiceProtocol {
         return medias
     }
     
-    func delete(media: MediaEntity) throws {
+    func delete(id: UUID) throws {
+        guard let media = try? self.context.fetch(MediaEntity.fetchRequest()).first(where: { $0.id == id }) else { return }
+        
         self.context.delete(media)
         try self.context.save()
     }
     
-    func move(media: MediaEntity, to album: AlbumEntity) throws {
+    func move(id: UUID, to album: AlbumEntity) throws {
+        guard let media = try? self.context.fetch(MediaEntity.fetchRequest()).first(where: { $0.id == id }) else { return }
+        
         media.album = album
         media.date_added = Date()
         try self.context.save()
     }
     
-    
+    func fetchFavorites() -> [MediaEntity] {
+        let request = MediaEntity.fetchRequest()
+        request.predicate = NSPredicate(format: "is_favorited == YES")
+        return (try? context.fetch(request)) ?? []
+    }
 }
